@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.2.3 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -39,9 +39,9 @@ dependencies:
   provider: ^6.1.1
 
   # Firebase
-  firebase_core: 2.32.0
-  firebase_auth: 4.15.3
-  cloud_firestore: 4.13.6
+  firebase_core: ^2.24.2
+  firebase_auth: ^4.15.3
+  cloud_firestore: ^4.13.6
   google_sign_in: 6.1.6
   firebase_core_web: 2.17.0
   firebase_auth_web: 5.12.0
