@@ -1,136 +1,82 @@
 import 'package:flutter/material.dart';
-import 'package:film_match_app/providers/movie_provider.dart';
-import 'package:cached_network_image/cached_network_image.dart';
+import '../models/movie.dart';
 
 class MovieCard extends StatelessWidget {
   final Movie movie;
-  final VoidCallback onLike;
-  final VoidCallback onDislike;
-  final bool isLiked;
+  final VoidCallback? onTap;
+  final VoidCallback? onWishlistToggle;
+  final bool isInWishlist;
 
   const MovieCard({
     super.key,
     required this.movie,
-    required this.onLike,
-    required this.onDislike,
-    required this.isLiked,
+    this.onTap,
+    this.onWishlistToggle,
+    this.isInWishlist = false,
   });
 
   @override
   Widget build(BuildContext context) {
     return Card(
-      margin: const EdgeInsets.only(bottom: 16),
       clipBehavior: Clip.antiAlias,
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          // Movie poster
-          AspectRatio(
-            aspectRatio: 16 / 9,
-            child: CachedNetworkImage(
-              imageUrl: movie.posterPath,
-              fit: BoxFit.cover,
-              placeholder: (context, url) => Container(
-                color: Colors.grey[300],
-                child: const Center(
-                  child: CircularProgressIndicator(),
-                ),
-              ),
-              errorWidget: (context, url, error) => Container(
-                color: Colors.grey[300],
-                child: const Icon(
-                  Icons.error,
-                  color: Colors.red,
-                ),
-              ),
-            ),
-          ),
-          Padding(
-            padding: const EdgeInsets.all(16),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                // Movie title and rating
-                Row(
-                  children: [
-                    Expanded(
-                      child: Text(
-                        movie.title,
-                        style: const TextStyle(
-                          fontSize: 16,
-                          fontWeight: FontWeight.bold,
-                        ),
-                        textAlign: TextAlign.center,
-                        maxLines: 2,
-                        overflow: TextOverflow.ellipsis,
+      child: InkWell(
+        onTap: onTap,
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            AspectRatio(
+              aspectRatio: 2 / 3,
+              child: movie.poster.isNotEmpty
+                  ? Image.network(
+                      movie.poster,
+                      fit: BoxFit.cover,
+                      errorBuilder: (context, error, stackTrace) {
+                        return const Center(
+                          child: Icon(
+                            Icons.movie,
+                            size: 50,
+                            color: Colors.grey,
+                          ),
+                        );
+                      },
+                    )
+                  : const Center(
+                      child: Icon(
+                        Icons.movie,
+                        size: 50,
+                        color: Colors.grey,
                       ),
                     ),
-                  ],
-                ),
-                const SizedBox(height: 5),
-                Text(
-                  'Год: ${movie.year}',
-                  style: const TextStyle(
-                    fontSize: 14,
-                    color: Colors.grey,
+            ),
+            Padding(
+              padding: const EdgeInsets.all(8.0),
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Text(
+                    movie.title,
+                    style: Theme.of(context).textTheme.titleMedium,
+                    maxLines: 2,
+                    overflow: TextOverflow.ellipsis,
                   ),
-                ),
-                const SizedBox(height: 10),
-                // Описание (overview) и рейтинг (voteAverage) больше не доступны напрямую из OMDb search results
-                // Вы можете получить их, сделав дополнительный запрос по IMDB ID, если это необходимо.
-                // Однако, для текущей цели отображения в карточке, давайте их удалим.
-                /*
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    const Icon(Icons.star, color: Colors.amber, size: 16),
-                    const SizedBox(width: 5),
-                    Text(
-                      movie.voteAverage.toStringAsFixed(1),
-                      style: const TextStyle(fontSize: 14),
-                    ),
-                  ],
-                ),
-                const SizedBox(height: 10),
-                Text(
-                  movie.overview,
-                  style: const TextStyle(fontSize: 12),
-                  maxLines: 3,
-                  overflow: TextOverflow.ellipsis,
-                  textAlign: TextAlign.center,
-                ),
-                */
-                const SizedBox(height: 16),
-                // Action buttons
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                  children: [
-                    // Dislike button
-                    ElevatedButton.icon(
-                      onPressed: onDislike,
-                      icon: const Icon(Icons.close),
-                      label: const Text('Пропустить'),
-                      style: ElevatedButton.styleFrom(
-                        backgroundColor: Colors.red,
-                        foregroundColor: Colors.white,
-                      ),
-                    ),
-                    // Like button
-                    ElevatedButton.icon(
-                      onPressed: onLike,
-                      icon: Icon(isLiked ? Icons.favorite : Icons.favorite_border),
-                      label: Text(isLiked ? 'В избранном' : 'Нравится'),
-                      style: ElevatedButton.styleFrom(
-                        backgroundColor: isLiked ? Colors.orange : Colors.green,
-                        foregroundColor: Colors.white,
-                      ),
-                    ),
-                  ],
-                ),
-              ],
+                  const SizedBox(height: 4),
+                  Text(
+                    movie.year,
+                    style: Theme.of(context).textTheme.bodySmall,
+                  ),
+                ],
+              ),
             ),
-          ),
-        ],
+            if (onWishlistToggle != null)
+              IconButton(
+                icon: Icon(
+                  isInWishlist ? Icons.favorite : Icons.favorite_border,
+                  color: isInWishlist ? Colors.red : null,
+                ),
+                onPressed: onWishlistToggle,
+              ),
+          ],
+        ),
       ),
     );
   }
